+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.0a4 (unreleased)
 ------------------
 
+- Add policy valiation for professional
+  [vpiret]
+
 - Add faceted for professional
   [vpiret]